--- app/src/main/res/layout/activity_auth.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:paddingHorizontal="18dp"
-    android:paddingTop="14dp"
-    android:paddingBottom="16dp">
-
-    <Button
-        android:id="@+id/btn_auth_start"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:paddingVertical="20dp"
-        tools:text="시작하기"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
